+++ macgyver-plugin-cloud-aws/src/main/java/io/macgyver/plugin/cloud/aws/AWSServiceClientImpl.java
@@ -15,7 +15,13 @@ package io.macgyver.plugin.cloud.aws;
 
 import java.lang.reflect.Constructor;
 import java.lang.reflect.InvocationTargetException;
-import java.util.Optional;
+import java.util.List;
+
+import org.lendingclub.mercator.aws.AWSScannerBuilder;
+import org.lendingclub.mercator.aws.AllEntityScanner;
+import org.lendingclub.mercator.core.Projector;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import com.amazonaws.AmazonWebServiceClient;
 import com.amazonaws.auth.AWSCredentialsProvider;
@@ -23,15 +29,21 @@ import com.amazonaws.regions.Region;
 import com.amazonaws.regions.Regions;
 import com.amazonaws.services.ec2.AmazonEC2Client;
 import com.amazonaws.services.s3.AmazonS3Client;
+import com.google.common.base.MoreObjects;
+import com.google.common.collect.ImmutableList;
+import com.google.common.collect.Lists;
 
+import io.macgyver.core.Kernel;
 import io.macgyver.core.MacGyverException;
 
 public class AWSServiceClientImpl implements AWSServiceClient {
 
-	AWSCredentialsProvider credentialsProvider;
-	Region defaultRegion=null;
-	String accountId;
+	Logger logger = LoggerFactory.getLogger(AWSServiceClientImpl.class);
 	
+	AWSCredentialsProvider credentialsProvider;
+
+	String accountId;
+	List<Regions> regionList = ImmutableList.of();
 	public  AWSServiceClientImpl() {
 		super();
 	}
@@ -50,12 +62,12 @@ public class AWSServiceClientImpl implements AWSServiceClient {
 
 	@Override
 	public AmazonS3Client newS3Client() {
-		return createClient(AmazonS3Client.class,defaultRegion);
+		return createClient(AmazonS3Client.class);
 	}
 
 	@Override
 	public AmazonEC2Client createEC2Client() {
-		return createClient(AmazonEC2Client.class,defaultRegion);
+		return createClient(AmazonEC2Client.class);
 	}
 
 	@Override
@@ -91,12 +103,9 @@ public class AWSServiceClientImpl implements AWSServiceClient {
 	public <T extends AmazonWebServiceClient> T createClient(
 			Class<? extends T> t) {
 
-		return createClient(t, defaultRegion);
+		return createClient(t,null);
 	}
 
-	public Optional<Region> getDefaultRegion() {
-		return Optional.ofNullable(defaultRegion);
-	}
 	
 	@Override
 	public <T extends AmazonWebServiceClient> T createClient(
@@ -128,4 +137,57 @@ public class AWSServiceClientImpl implements AWSServiceClient {
 		this.accountId = id;
 	}
 
+	public AWSScannerBuilder createScannerBuilder() {
+		// No need to set region or account.  Account will be determined at runtime.  Region will be selected by the caller.
+		return Kernel.getApplicationContext().getBean(Projector.class).createBuilder(AWSScannerBuilder.class).withCredentials(getCredentialsProvider()).withFailOnError(false);
+	}
+
+	@Override
+	public void scanRegion(String name) {
+		scanRegion(Regions.fromName(name));	
+	}
+
+	@Override
+	public void scanRegion(Regions region) {		
+		createScannerBuilder().withRegion(region).build(AllEntityScanner.class).scan();
+	}
+
+	@Override
+	public void scanRegion(Region region) {
+		scanRegion(Regions.fromName(region.getName()));
+	}
+
+	public void scan() {
+		for (Regions region: getConfiguredRegions()) {
+			try {
+				scanRegion(region);
+			}
+			catch (RuntimeException e) {
+				logger.warn("problem scanning region: "+region,e);
+			}
+		}
+	}
+	
+	/**
+	 * This is public, but only on the implementation class.
+	 * @param list
+	 */
+	
+	public synchronized void setConfiguredRegions(List<Regions> list) {
+		if (list==null) {
+			list = Lists.newArrayList();
+		}
+		regionList = ImmutableList.copyOf(list);
+	}
+	
+	@Override
+	public synchronized List<Regions> getConfiguredRegions() {
+		return regionList;
+	}
+	
+	public String toString() {
+		return MoreObjects.toStringHelper(this).add("account", getAccountId()).toString();
+		
+	}
+	
 }